refactor(park-keeper): 清除死碼與整併重複 MiniMap 測試
- 刪除 routes 未引用的 pages/Settings.tsx 與其測試（/settings 路由直用 Home initialTab）
- 刪除閒置的 components/HomeSkeleton.tsx 與其測試
- 刪除僅測試引用的 services/geolocation.ts 與其測試（生產碼直用 navigator.geolocation）
- 完整版 MiniMap 測試搬移至 components/__tests__/，整併原重複檔 4 個照片拖曳測試，斷言零刪除
- 測試數 335→320（-15：Settings 1、HomeSkeleton 2、geolocation 12），MiniMap 58 個全保留

測試：pnpm --filter @app/park-keeper typecheck、test --run（27 files / 320 tests）、build 全綠
